Integrated Extent report
