--- SeleniumFramework/src/test/java/org/selenium/pojoObjects/Product.java
@@ -1,40 +1,0 @@
-package org.selenium.pojoObjects;
-
-import org.selenium.utils.JacksonUtils;
-
-import java.io.IOException;
-
-public class Product {
-    private int id;
-    private String name;
-
-    public Product(int id) throws IOException {
-        Product[] products = JacksonUtils.deserializeJson("products.json", Product[].class);
-        for(Product product:products){
-            if(product.getId()==id){
-                this.id = id;
-                this.name = product.getName();
-            }
-        }
-    }
-
-    public Product(){
-
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
