--- SeleniumFramework/src/test/java/org/selenium/utils/JacksonUtils.java
@@ -1,17 +1,0 @@
-package org.selenium.utils;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.selenium.pojoObjects.BillingAddress;
-
-import java.io.IOException;
-import java.io.InputStream;
-
-public class JacksonUtils {
-
-    public static <T> T deserializeJson(String fileName, Class<T> T) throws IOException {
-        InputStream is =JacksonUtils.class.getClassLoader().getResourceAsStream(fileName);
-        ObjectMapper objectMapper = new ObjectMapper();
-        return objectMapper.readValue(is, T);
-    }
-}
