--- SeleniumFramework/testng.xml
@@ -1,17 +1,0 @@
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
-
-<suite name="Suite1" verbose="1" thread-count="2" parallel="tests">
-    <test name="Chrome" thread-count="2" parallel="methods">
-        <parameter name="browser" value="CHROME"></parameter>
-        <classes>
-            <class name="org.selenium.tests.FirstTestCase" />
-        </classes>
-    </test>
-
-    <!--<test name="Firefox" thread-count="2" parallel="methods">
-        <parameter name="browser" value="FIREFOX"></parameter>
-        <classes>
-            <class name="org.selenium.tests.FirstTestCase"/>
-        </classes>
-    </test>-->
-</suite>
