+++ src/test/java/org/selenium/utils/JacksonUtils.java
@@ -1,0 +1,15 @@
+package org.selenium.utils;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+import java.io.IOException;
+import java.io.InputStream;
+
+public class JacksonUtils {
+
+    public static <T> T deserializeJson(String fileName, Class<T> T) throws IOException {
+        InputStream is = JacksonUtils.class.getClassLoader().getResourceAsStream(fileName);
+        ObjectMapper objectMapper = new ObjectMapper();
+        return objectMapper.readValue(is, T);
+    }
+}
